Move sampling to own directory, reduce duplicates
